Amended after code review
