+++ Threads/task7/src/main/java/com/epam/concurrency/run/Producer.java
@@ -5,26 +5,13 @@ import org.apache.log4j.Logger;
 /**
  * Created by Andrey Yun on 13.02.2016.
  */
-public class Producer implements Runnable {
+public class Producer extends ResourceUser {
 
     private static final Logger LOG = Logger.getLogger(Producer.class);
 
-    private CommonResource resource;
-    private static int counter;
-    private final int ID = ++counter;
-
-    public CommonResource getResource() {
-        return resource;
-    }
-
-    public void setResource(CommonResource resource) {
-        this.resource = resource;
-    }
-
-    public void run() {
-        while (true) {
-            LOG.info("Producer " + Thread.currentThread().getName() + " after update: " +
-                    resource.incrementCounterIfLessThanTen());
-        }
+    @Override
+    protected void doAction() {
+        LOG.info("Producer " + Thread.currentThread().getName() + " after update: " +
+                getResource().incrementCounterIfLessThanTen());
     }
 }
