+++ Threads/task7/src/main/java/com/epam/concurrency/run/Runner.java
@@ -14,22 +14,17 @@ public class Runner {
 
     private static final Logger LOG = Logger.getLogger(Runner.class);
 
-    private static final int NUMBER_OF_CONSUMERS_PRODUCERS = 10;
+    private static final int NUMBER_OF_CONSUMERS_PRODUCERS = 5;
 
     public static void main(String[] args) {
         ApplicationContext ctx = new AnnotationConfigApplicationContext(SpringConfig.class);
         ExecutorService exex = Executors.newCachedThreadPool();
-        try {
-            for (int i = 0; i < NUMBER_OF_CONSUMERS_PRODUCERS; i++) {
-                Consumer consumer = (Consumer) ctx.getBean(SpringConfig.CONSUMER_BEAN_NAME);
-                Producer producer = (Producer) ctx.getBean(SpringConfig.PRODUCER_BEAN_NAME);
-                exex.execute(consumer);
-                exex.execute(producer);
-            }
-        } catch (Exception e) {
-            LOG.error(e.getMessage(), e);
-        } finally {
-            exex.shutdown();
+        for (int i = 0; i < NUMBER_OF_CONSUMERS_PRODUCERS; i++) {
+            Consumer consumer = (Consumer) ctx.getBean(SpringConfig.CONSUMER_BEAN_NAME);
+            Producer producer = (Producer) ctx.getBean(SpringConfig.PRODUCER_BEAN_NAME);
+            exex.execute(consumer);
+            exex.execute(producer);
         }
+        exex.shutdown();
     }
 }
